pyproject.toml
[tool.poetry]
name = "qb-bot"
version = "1.0.0-rc.0"
description = "quizbowl discord bot"
authors = ["Sky \"g3ner1c\" Hong <g3ner1c-sky@pm.me>"]
license = "AGPL-3.0-or-later"
readme = "README.md"

repository = "https://www.github.com/g3ner1c/qb-bot"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.1"
markdownify = "^0.11"
aiohttp = "^3.8"
aiodns = "^3.0"


[tool.poetry.dev-dependencies]
flake8 = "^6.0"
isort = "^5.10"
black = "^23.1"
pydocstyle = "^6.3"
pre-commit = "^2.20"
python-dotenv = "^0.21"

[tool.black]
line-length = 99

[tool.isort]
profile = "black"
line_length = 99

[tool.flake8]
max_line_length = 99
exclude = ".vscode,.git,__pycache__,.venv"

[tool.pydocstyle]
convention = "numpy"
